Data_Viz assignment 1
This is a basic html webpage with pictures and links to five articles I've used in my research. The stories talk about Burmese refugees throughout the world. A large number are in Thailand and there was an attempt last year for repatriation. The links from these stories are from The Diplomat, Time Magazine, UNHCR and Al Jazeera.